backend/requirements.txt
annotated-doc==0.0.3; python_version >= "3.8"
annotated-types==0.7.0; python_version >= "3.9"
anyio==4.11.0; python_version >= "3.9"
certifi==2025.10.5; python_version >= "3.8"
click==8.3.0; python_version >= "3.10"
colorama==0.4.6; python_version >= "3.10" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.10" and python_full_version >= "3.7.0" and platform_system == "Windows"
exceptiongroup==1.3.0; python_version < "3.11" and python_version >= "3.9"
fastapi==0.120.4; python_version >= "3.8"
h11==0.16.0; python_version >= "3.9"
httpcore==1.0.9; python_version >= "3.8"
httpx==0.28.1; python_version >= "3.8"
idna==3.11; python_version >= "3.9"
iniconfig==2.3.0; python_version >= "3.10"
packaging==25.0; python_version >= "3.9"
pluggy==1.6.0; python_version >= "3.9"
pydantic-core==2.41.4; python_version >= "3.9"
pydantic-settings==2.11.0; python_version >= "3.9"
pydantic==2.12.3; python_version >= "3.9"
pygments==2.19.2; python_version >= "3.9"
pytest==8.4.2; python_version >= "3.9"
python-dotenv==1.2.1; python_version >= "3.9"
sniffio==1.3.1; python_version >= "3.9"
starlette==0.49.3; python_version >= "3.9"
tomli==2.3.0; python_version < "3.11" and python_version >= "3.9"
typing-extensions==4.15.0; python_version < "3.11" and python_version >= "3.9"
typing-inspection==0.4.2; python_version >= "3.9"
uvicorn==0.38.0; python_version >= "3.9"
